+++ khorosjx/news.py
@@ -6,7 +6,7 @@
 :Example:        ``all_publication = khorosjx.news.get_all_publications()``
 :Created By:     Jeff Shurtliff
 :Last Modified:  Jeff Shurtliff
-:Modified Date:  10 Feb 2020
+:Modified Date:  11 Feb 2020
 """
 
 from . import core, errors
@@ -92,6 +92,10 @@ def get_publication(pub_id, return_fields=[], ignore_exceptions=False):
     :returns: A dictionary with the data for the publication
     :raises: InvalidDatasetError, GETRequestError
     """
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Retrieve the publication
     publication = core.get_data('publications', pub_id, return_json=True, all_fields=True)
     successful_response = errors.handlers.check_api_response(publication, ignore_exceptions=ignore_exceptions)
     if successful_response:
@@ -108,6 +112,10 @@ def delete_publication(pub_id, return_json=False):
     :type return_json: bool
     :returns: The API response (optionally in JSON format)
     """
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Delete the publication
     publication_uri = f"{base_url}/publications/{pub_id}"
     response = core.delete(publication_uri, return_json=return_json)
     return response
@@ -132,6 +140,10 @@ def get_subscription_ids(pub_id, return_type='str'):
     :returns: A list of subscription IDs
     :raises: ValueError
     """
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Retrieve the subscription IDs
     subscription_ids = []
     all_subscriptions = get_subscription_data(pub_id)
     for subscription in all_subscriptions:
@@ -157,10 +169,14 @@ def get_subscriber_groups(publication_id, subscription_id='', full_uri=False):
     :raises: SubscriptionNotFoundError
     """
     def __filter_subscriptions_by_id(_sub_id, _subscriptions):
+        """This function filters the returned IDs by a supplied subscription ID when applicable."""
         for _subscription in _subscriptions:
             if _subscription['id'] == _sub_id:
                 return _subscription
         raise errors.exceptions.SubscriptionNotFoundError
+
+    # Verify that the core connection has been established
+    verify_core_connection()
 
     # Capture the subscriber groups for each subscription
     subscriptions = get_subscription_data(publication_id)
@@ -182,8 +198,98 @@ def get_subscriber_groups(publication_id, subscription_id='', full_uri=False):
     return subscriber_groups
 
 
-def get_subscribers(publication_id, subscription_id='', full_uri=False):
-    # TODO: https://developers.jivesoftware.com/api/v3/cloud/rest/PublicationService.html#getSubscriberUsers(long,%20long,%20int,%20int,%20String)
-    pass
+def get_subscribers(publication_id, subscription_id, return_type='list', only_id=True, return_fields=[],
+                    ignore_exceptions=False):
+    """This function retrieves the individual subscribers (i.e. users) for a given subscription within a publication.
+
+    :param publication_id: The ID of the publication where the subscription resides
+    :type publication_id: int, str
+    :param subscription_id: The ID of the subscription in which to identify the subscribers
+    :type subscription_id: int, str
+    :param return_type: Determines whether the data should be returned as a ``list`` (default) or a pandas ``dataframe``
+    :type return_type: str
+    :param only_id: Determines if only the ID of each user should be returned (default) or a dict with all user data
+    :type only_id: bool
+    :param return_fields: Specific fields to return if not all of the default fields are needed (Optional)
+    :type return_fields: list
+    :param ignore_exceptions: Determines whether nor not exceptions should be ignored (Default: ``False``)
+    :type ignore_exceptions: bool
+    :returns: A list or pandas dataframe with the subscriber information
+    """
+    def __get_subscriber_ids(_subscribers):
+        """This function pulls the subscriber IDs out of dictionaries and into a single list."""
+        _subscriber_ids = []
+        for _subscriber in _subscribers:
+            _subscriber_ids.append(_subscriber['id'])
+        return _subscriber_ids
+
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Initialize the empty list for the subscription information
+    all_subscribers = []
+
+    # Overwrite the return_fields list if the only_id value is True
+    if only_id:
+        return_fields = ['id']
+
+    # Perform the first query to get up to the first 100 groups
+    query = f"{base_url}/publications/{publication_id}/subscriptions/{subscription_id}/subscribers"
+    start_index = 0
+    subscribers = core.get_paginated_results(query, 'people', start_index, return_fields=return_fields,
+                                             ignore_exceptions=ignore_exceptions)
+    if only_id and return_type == 'list':
+        subscribers = __get_subscriber_ids(subscribers)
+    all_subscribers = core_utils.add_to_master_list(subscribers, all_subscribers)
+
+    # Continue querying for groups until none are returned
+    while len(subscribers) > 0:
+        start_index += 100
+        subscribers = core.get_paginated_results(query, 'people', start_index, return_fields=return_fields,
+                                                 ignore_exceptions=ignore_exceptions)
+        if only_id and return_type == 'list':
+            subscribers = __get_subscriber_ids(subscribers)
+        all_subscribers = core_utils.add_to_master_list(subscribers, all_subscribers)
+
+    # Return the data as a master list of publication dictionaries or a pandas dataframe
+    if return_type == "dataframe":
+        all_subscribers = df_utils.convert_dict_list_to_dataframe(all_subscribers)
+    return all_subscribers
 
 
+def rebuild_publication(publication_id):
+    """This function rebuilds a publication.
+
+    :param publication_id: The ID of the publication to be rebuilt
+    :type publication_id: int, str
+    :returns: The response from the API PUT request
+    """
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Perform the PUT request to rebuild the publication
+    query = f"{base_url}/publications/{publication_id}/rebuild"
+    payload = {}
+    response = core.put_request_with_retries(query, payload)
+    return response
+
+
+def update_publication(publication_id, payload):
+    """This function updates a publication using the supplied JSON payload.
+
+    :param publication_id: The ID of the publication to be updated
+    :type publication_id: int, str
+    :param payload: The JSON payload with which the publication will be updated
+    :type payload: dict
+    :returns: The response from the API PUT request
+    """
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Perform the PUT request to update the publication
+    query = f"{base_url}/publications/{publication_id}"
+    response = core.put_request_with_retries(query, payload)
+    return response
+
+
+# TODO: Add functions relating to the /streams and /people/{id}/streams endpoints